# docs/openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  title: Gen3Workflow
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /s3/{path}:
    delete:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    get:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    head:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    options:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    patch:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    post:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    put:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    trace:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
  /storage/info:
    get:
      operationId: get_storage_info
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Get Storage Info Storage Info Get
                type: object
          description: Successful Response
      security:
      - HTTPBearer: []
      summary: Get Storage Info
      tags:
      - Storage
...
  /{path}:
    delete:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    get:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    head:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    options:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    patch:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    post:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    put:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3
    trace:
      description: 'Receive incoming signed S3 requests, re-sign them (AWS Signature
        Version 4 algorithm) with the

        appropriate credentials to access the current user''s AWS S3 bucket, and forward
        them to

        AWS S3. The recommended way to use this endpoint is to use the AWS SDK or
        CLI.


        The S3 endpoint is exposed at `/s3` as well as it the root `/` because S3
        clients like Minio do

        not support S3 endpoints with a path.'
      operationId: s3_endpoint_2
      parameters:
      - in: path
        name: path
        required: true
        schema:
          title: Path
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: S3 Endpoint
      tags:
      - S3